--- feature/music_player/src/main/java/com/example/feature/music_player/data/remote/MusicSource.kt
@@ -1,8 +1,0 @@
-package com.example.musicplayer.data.remote
-
-import com.example.musicplayer.data.entities.Song
-
-interface MusicSource {
-
-    suspend fun getMusic(): List<Song>
-}
