+++ feature/music_player/src/main/java/com/example/feature/music_player/data/source/MusicSource.kt
@@ -1,0 +1,8 @@
+package com.example.feature.music_player.data.source
+
+import com.example.feature.music_player.data.entities.Song
+
+interface MusicSource {
+
+    suspend fun getMusic(): List<Song>
+}
