+++ feature/music_player/src/main/java/com/example/feature/music_player/di/AppModule.kt
@@ -5,7 +5,7 @@ import com.bumptech.glide.Glide
 import com.bumptech.glide.load.engine.DiskCacheStrategy
 import com.bumptech.glide.request.RequestOptions
 import com.example.feature.music_player.R
-import com.example.musicplayer.exoplayer.MusicServiceConnection
+import com.example.feature.music_player.exoplayer.MusicServiceConnection
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
