+++ feature/music_player/src/main/java/com/example/feature/music_player/exoplayer/callbacks/MusicPlaybackPrepared.kt
@@ -1,11 +1,11 @@
-package com.example.musicplayer.exoplayer.callbacks
+package com.example.feature.music_player.exoplayer.callbacks
 
 import android.net.Uri
 import android.os.Bundle
 import android.os.ResultReceiver
 import android.support.v4.media.MediaMetadataCompat
 import android.support.v4.media.session.PlaybackStateCompat
-import com.example.musicplayer.exoplayer.music_source.MusicProvider
+import com.example.feature.music_player.exoplayer.music_source.MusicProvider
 import com.google.android.exoplayer2.Player
 import com.google.android.exoplayer2.ext.mediasession.MediaSessionConnector
