+++ feature/music_player/src/main/java/com/example/feature/music_player/extensions/CombineExt.kt
@@ -1,29 +1,9 @@
 @file:Suppress("UNCHECKED_CAST")
 
-package com.example.musicplayer.extensions
+package com.example.feature.music_player.extensions
 
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.combine as coroutinesFlowCombine
-
-fun <T1, T2, T3, T4, T5, T6, R> combine(
-    flow: Flow<T1>,
-    flow2: Flow<T2>,
-    flow3: Flow<T3>,
-    flow4: Flow<T4>,
-    flow5: Flow<T5>,
-    flow6: Flow<T6>,
-    transform: suspend (T1, T2, T3, T4, T5, T6) -> R
-): Flow<R> =
-    coroutinesFlowCombine(flow, flow2, flow3, flow4, flow5, flow6) { args: Array<*> ->
-        transform(
-            args[0] as T1,
-            args[1] as T2,
-            args[2] as T3,
-            args[3] as T4,
-            args[4] as T5,
-            args[5] as T6,
-        )
-    }
 
 fun <T1, T2, T3, T4, T5, T6, T7, R> combine(
     flow: Flow<T1>,
