+++ feature/music_player/src/main/java/com/example/feature/music_player/ui/composables/music_player_widget/MusicPlayerWidget.kt
@@ -1,4 +1,4 @@
-package com.example.musicplayer.ui.composables.music_player_widget
+package com.example.feature.music_player.ui.composables.music_player_widget
 
 
 import androidx.compose.animation.AnimatedVisibility
@@ -35,9 +35,9 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.IntOffset
 import androidx.compose.ui.unit.dp
 import com.example.feature.music_player.R
+import com.example.feature.music_player.data.entities.MediaDataSource
+import com.example.feature.music_player.data.entities.Song
 import com.example.feature.music_player.ui.viewmodels.MusicAction
-import com.example.musicplayer.data.entities.MediaDataSource
-import com.example.musicplayer.data.entities.Song
 import com.example.musicplayer.ui.theme.MusicPlayerTheme
 import com.google.accompanist.coil.rememberCoilPainter
 import kotlin.math.roundToInt
