+++ feature/music_player/src/main/java/com/example/feature/music_player/ui/music_player/composables/MusicControlPanel.kt
@@ -23,9 +23,9 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.example.feature.music_player.R
+import com.example.feature.music_player.data.entities.MediaDataSource
+import com.example.feature.music_player.data.entities.Song
 import com.example.feature.music_player.ui.viewmodels.MusicAction
-import com.example.musicplayer.data.entities.MediaDataSource
-import com.example.musicplayer.data.entities.Song
 import com.example.musicplayer.ui.theme.MusicPlayerTheme
 
 @Composable
