+++ feature/music_player/src/main/java/com/example/feature/music_player/ui/viewmodels/Action.kt
@@ -1,6 +1,6 @@
 package com.example.feature.music_player.ui.viewmodels
 
-import com.example.musicplayer.data.entities.Song
+import com.example.feature.music_player.data.entities.Song
 
 sealed interface MusicAction {
